Don't legalise the perms or mode in the fetch stage
This was causing an assertion failure as non-control flow instructions
would change the PC in speculation when they legalised the mode bit.
Removed the comments as they're technically not true. Justification
that using the unlegalised version is okay would be more like
"If we will commit, then the PC must be legal"